move GA code to the main app entry
