--- dotcom-rendering/src/web/browser/ga/init.ts
@@ -1,12 +1,0 @@
-import '../webpackPublicPath';
-import { startup } from '../startup';
-import { init as initGa, sendPageView } from './ga';
-
-const init = (): Promise<void> => {
-	initGa();
-	sendPageView();
-
-	return Promise.resolve();
-};
-
-startup('ga', null, init);
